working now

+++ public/src/components/logout.jsx
@@ -1,23 +1,26 @@
-import React from "react"
-import { useNavigate } from "react-router-dom"
-import { BiPowerOff } from "react-icons/bi"
-import styled from "styled-components"
-import axios from "axios"
-import { logoutRoute } from "../utils/APIRoutes"
-
+import React from "react";
+import { useNavigate } from "react-router-dom";
+import { BiPowerOff } from "react-icons/bi";
+import styled from "styled-components";
+import axios from "axios";
+import { logoutRoute } from "../utils/APIRoutes";
 export default function Logout() {
-    const navigate = useNavigate()
-
-    const handleClick = async () => {
-        localStorage.clear()
-      navigate("/login")
+  const navigate = useNavigate();
+  const handleClick = async () => {
+    const id = await JSON.parse(
+      localStorage.getItem(process.env.REACT_APP_LOCALHOST_KEY)
+    )._id;
+    const data = await axios.get(`${logoutRoute}/${id}`);
+    if (data.status === 200) {
+      localStorage.clear();
+      navigate("/login");
     }
-
-    return (
-        <Button onClick={handleClick}>
-          <BiPowerOff />
-        </Button>
-      )
+  };
+  return (
+    <Button onClick={handleClick}>
+      <BiPowerOff />
+    </Button>
+  );
 }
 
 const Button = styled.button`
@@ -33,4 +36,4 @@ const Button = styled.button`
     font-size: 1.3rem;
     color: #ebe7ff;
   }
-`
+`;
